+++ deicode/q2/tests/test_method.py
@@ -8,9 +8,8 @@ from skbio.stats.distance import DistanceMatrix
 from skbio.util import get_data_path
 from qiime2 import Artifact
 from qiime2.plugins import deicode as q2deicode
-from deicode.rpca import rpca
-from deicode.scripts._standalone_rpca import standalone_rpca
-from deicode.testing import assert_deicode_ordinationresults_equal
+from deicode.rpca import rpca, auto_rpca
+from deicode.scripts._standalone_rpca import deicode as sdc
 from simulations import build_block_model
 from click.testing import CliRunner
 from nose.tools import nottest
@@ -50,6 +49,18 @@ class Testrpca(unittest.TestCase):
         self.assertFalse(np.isnan(ord_test.features).any(axis=None))
         self.assertFalse(np.isnan(ord_test.samples).any(axis=None))
 
+    def test_auto_rpca(self):
+        """Tests the basic validity of the actual auto_rpca()."""
+        ord_test, dist_test = auto_rpca(table=self.test_table)
+        # Validate types of the RPCA outputs
+        self.assertIsInstance(ord_test, OrdinationResults)
+        self.assertIsInstance(dist_test, DistanceMatrix)
+        # Ensure that no NaNs are in the OrdinationResults
+        # NOTE that we have to use the DataFrame .any() functions instead of
+        # python's built-in any() functions -- see #29 for details on this
+        self.assertFalse(np.isnan(ord_test.features).any(axis=None))
+        self.assertFalse(np.isnan(ord_test.samples).any(axis=None))
+
 
 class Test_qiime2_rpca(unittest.TestCase):
 
@@ -57,14 +68,15 @@ class Test_qiime2_rpca(unittest.TestCase):
         self.q2table = Artifact.import_data("FeatureTable[Frequency]",
                                             create_test_table())
 
-    def test_qiime2_rpca(self):
-        """Tests that the Q2 and standalone RPCA results match."""
+    def test_qiime2_auto_rpca(self):
+        """ Test Q2 rank estimate matches standalone."""
 
         tstdir = "test_output"
         # Run DEICODE through QIIME 2 (specifically, the Artifact API)
-        ordination_qza, distmatrix_qza = q2deicode.actions.rpca(self.q2table)
+        res = q2deicode.actions.auto_rpca(self.q2table)
+        ordination_qza, distmatrix_qza = res
         # Get the underlying data from these artifacts
-        q2ordination = ordination_qza.view(OrdinationResults)
+        # q2ordination = ordination_qza.view(OrdinationResults)
         q2distmatrix = distmatrix_qza.view(DistanceMatrix)
 
         # Next, run DEICODE outside of QIIME 2. We're gonna check that
@@ -77,12 +89,56 @@ class Test_qiime2_rpca(unittest.TestCase):
         tstdir_absolute = os_path_sep.join(q2table_loc.split(os_path_sep)[:-1])
 
         # Run DEICODE outside of QIIME 2...
-        CliRunner().invoke(standalone_rpca, ['--in-biom', q2table_loc,
-                                             '--output-dir', tstdir_absolute])
+        CliRunner().invoke(sdc.commands['auto-rpca'],
+                           ['--in-biom', q2table_loc,
+                           '--output-dir', tstdir_absolute])
         # ...and read in the resulting output files. This code was derived from
         # test_standalone_rpca() elsewhere in DEICODE's codebase.
-        stordination = OrdinationResults.read(get_data_path('ordination.txt',
-                                                            tstdir))
+        # stordination = OrdinationResults.read(get_data_path('ordination.txt',
+        #                                                    tstdir))
+        stdistmatrix_values = read_csv(
+            get_data_path(
+                'distance-matrix.tsv',
+                tstdir),
+            sep='\t',
+            index_col=0).values
+
+        # Convert the DistanceMatrix object a numpy array (which we can compare
+        # with the other _values numpy arrays we've created from the other
+        # distance matrices)
+        q2distmatrix_values = q2distmatrix.to_data_frame().values
+
+        # Finaly: actually check the consistency of Q2 and standalone results!
+        np.testing.assert_array_almost_equal(q2distmatrix_values,
+                                             stdistmatrix_values)
+
+    def test_qiime2_rpca(self):
+        """Tests that the Q2 and standalone RPCA results match."""
+
+        tstdir = "test_output"
+        # Run DEICODE through QIIME 2 (specifically, the Artifact API)
+        ordination_qza, distmatrix_qza = q2deicode.actions.rpca(self.q2table)
+        # Get the underlying data from these artifacts
+        # q2ordination = ordination_qza.view(OrdinationResults)
+        q2distmatrix = distmatrix_qza.view(DistanceMatrix)
+
+        # Next, run DEICODE outside of QIIME 2. We're gonna check that
+        # everything matches up.
+        # ...First, though, we need to write the contents of self.q2table to a
+        # BIOM file, so DEICODE can understand it.
+        self.q2table.export_data(get_data_path("", tstdir))
+        q2table_loc = get_data_path('feature-table.biom', tstdir)
+        # Derived from a line in test_standalone_rpca()
+        tstdir_absolute = os_path_sep.join(q2table_loc.split(os_path_sep)[:-1])
+
+        # Run DEICODE outside of QIIME 2...
+        CliRunner().invoke(sdc.commands['rpca'],
+                           ['--in-biom', q2table_loc,
+                            '--output-dir', tstdir_absolute])
+        # ...and read in the resulting output files. This code was derived from
+        # test_standalone_rpca() elsewhere in DEICODE's codebase.
+        # stordination = OrdinationResults.read(get_data_path('ordination.txt',
+        #                                                    tstdir))
         stdistmatrix_values = read_csv(
             get_data_path(
                 'distance-matrix.tsv',