+++ deicode/scripts/__init__.py
@@ -1,3 +1,0 @@
-from ._standalone_rpca import standalone_rpca
-
-__all__ = ["standalone_rpca"]