+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # DEICODE changelog
 
+## Version 0.2.4 (2020-01-07)
+
+### Features
+
+* Feature presence frequency filter [raised in issue #45](https://github.com/biocore/DEICODE/issues/45)
+
+This new feature allows users to choose a number from 0-100, with decimals allowed. The filter
+will remove features that are present in less than that percentage of samples. In practice
+this filter is useful to filter out features that will be difficult to use for log-ratios
+downstream in [Qurro](https://github.com/biocore/qurro).
+
+* 'optspace' option in n-components to make an estimation on the rank parameter. See below in bug-fixes for more info.
+
+### Bug fixes
+
+* Axis order [raised in issue #52](https://github.com/biocore/DEICODE/issues/52) and [issue #32](https://github.com/biocore/DEICODE/issues/32)
+
+This was partially fixed in PR #33 and PR #34 but there was a remaining bug
+in the optspace function [here](https://github.com/biocore/DEICODE/blob/b1f37059b79f6ca2e2db11ba2fb7500c1a92f87e/deicode/optspace.py#L211-L212).
+This bug was subtle and only caused the order to change periodically. Examples of this fix being tested on both
+real and simulated data can be found [here](https://nbviewer.jupyter.org/github/cameronmartino/hartig-net/blob/master/percent-explained/real-data-fraction-var.ipynb) and [here](https://nbviewer.jupyter.org/github/cameronmartino/hartig-net/blob/master/percent-explained/simulation-fraction-var.ipynb) respectively. 
+
+* Fraction total variance explained [raised in issue #53](https://github.com/biocore/DEICODE/issues/53)
+
+The fraction variance explained currently is calculated by the sum of squares based on the number of singular values. The number of singular values changes based on the rank input. This causes the resulting fraction variance explained to change dramatically as the rank is increased or decreased. This has been brought up multiple times on the QIIME2 forum. For example, see [here](). Additionally, the correct rank that actually explains the total variance is not clear (i.e. What rank should I choose?). To solve both of these problems a rank estimation is made based on part C of Keshavan, R. H., Montanari, A. & Oh, S. (Low-rank matrix completion with noisy observations: A quantitative comparison. in 2009 47th Annual Allerton Conference on Communication, Control, and Computing (Allerton) 1216–1222 (2009)) is used. This can can be enabled by setting n-components to 'optspace'. To prevent user confusion the default was not changed in this version. However, a future warning was added to warn users that in the next version 'optspace' based rank estimation will be the default.
+
 ## Version 0.2.3 (2019-6-18)
 
 ### Backward-incompatible changes [stable]